+++ src/app/dashboard/page.tsx
@@ -13,6 +13,7 @@ import ReturnDashboard from "./return/page";
 import StatusDashboard from "./status/page";
 import TransferDashboard from "./transfer/page";
 import { useEffect, useState } from "react";
+import { useAuth } from "@/components/providers";
 
 interface Medicine {
     ID: string;
@@ -28,34 +29,23 @@ interface Medicine {
 
 export default function Dashboard() {
     const router = useRouter();
+    const { user, loading } = useAuth(); // ใช้ context
     const [medicines, setMedicines] = useState([]);
     const [selectedTab, setSelectedTab] = useState('borrow');
     const [totalBorrow, setTotalBorrow] = useState(0);
     const [loggedInHospital, setLoggedInHospital] = useState('Na Mom Hospital');
-
-    // const [medicines, setMedicines] = useState<Medicine[]>([]);
-    const [user, setUser] = useState<{ email: string } | null>(null);
     const [isLoading, setIsLoading] = useState(true);
     const [error, setError] = useState<string | null>(null);
-    
-    useEffect(() => {
-        // Fetch user data
-        const fetchUser = async () => {
-            try {
-                const res = await fetch('/api/auth/me');
-                if (res.ok) {
-                    const data = await res.json();
-                    setUser(data);
-                } else {
-                    router.push('/');
-                }
-            } catch (error) {
-                console.error('Error fetching user:', error);
-                router.push('/');
-            }
-        };
 
-        // Fetch medicines data
+     // redirect ถ้าไม่ login
+    useEffect(() => {
+        if (!loading && !user) {
+            router.push('/');
+        }
+    }, [user, loading, router]);
+
+    // fetch medicines
+    useEffect(() => {
         const fetchMedicines = async () => {
             try {
                 setIsLoading(true);
@@ -65,42 +55,35 @@ export default function Dashboard() {
                     throw new Error('Failed to fetch medicines');
                 }
                 const data = await response.json();
-                
-                // Ensure data is an array
                 if (Array.isArray(data)) {
                     setMedicines(data);
                 } else if (data && typeof data === 'object') {
-                    // If data is an object, try to convert it to an array
-                    const medicinesArray = Object.values(data);
-                    setMedicines(medicinesArray);
+                    setMedicines(Object.values(data));
                 } else {
                     setMedicines([]);
                     setError('Invalid data format received');
                 }
             } catch (error) {
-                console.error('Error fetching medicines:', error);
                 setError('Failed to load medicines');
                 setMedicines([]);
             } finally {
                 setIsLoading(false);
             }
         };
-
-        fetchUser();
         fetchMedicines();
-    }, [router]);
+    }, []);
 
     useEffect(() => {
-        fetch("api/queryAll")
+        fetch("/api/queryAll")
             .then((response) => response.json())
             .then((data) => {
-                const filteredMedicines = data.filter(medicine => medicine.PostingHospital !== loggedInHospital);
+                const filteredMedicines = data.filter((medicine: any) => medicine.PostingHospital !== loggedInHospital);
                 setMedicines(filteredMedicines);
 
-                const borrowCount = data.reduce((total, medicine) => {
+                const borrowCount = data.reduce((total: any, medicine: any) => {
                     if (!medicine.BorrowRecords) return total;
                     const hospitalBorrows = medicine.BorrowRecords.filter(
-                        record => record.BorrowingHospital === loggedInHospital
+                        (record: any) => record.BorrowingHospital === loggedInHospital
                     ).length;
                     return total + hospitalBorrows;
                 }, 0);
@@ -109,6 +92,78 @@ export default function Dashboard() {
             })
             .catch((error) => console.error("Error fetching data:", error));
     }, [loggedInHospital]);
+    
+    // useEffect(() => {
+    //     // Fetch user data
+    //     const fetchUser = async () => {
+    //         try {
+    //             const res = await fetch('/api/auth/me');
+    //             if (res.ok) {
+    //                 const data = await res.json();
+    //                 setUser(data);
+    //             } else {
+    //                 router.push('/');
+    //             }
+    //         } catch (error) {
+    //             console.error('Error fetching user:', error);
+    //             router.push('/');
+    //         }
+    //     };
+
+    //     // Fetch medicines data
+    //     const fetchMedicines = async () => {
+    //         try {
+    //             setIsLoading(true);
+    //             setError(null);
+    //             const response = await fetch('/api/queryAll');
+    //             if (!response.ok) {
+    //                 throw new Error('Failed to fetch medicines');
+    //             }
+    //             const data = await response.json();
+                
+    //             // Ensure data is an array
+    //             if (Array.isArray(data)) {
+    //                 setMedicines(data);
+    //             } else if (data && typeof data === 'object') {
+    //                 // If data is an object, try to convert it to an array
+    //                 const medicinesArray = Object.values(data);
+    //                 setMedicines(medicinesArray);
+    //             } else {
+    //                 setMedicines([]);
+    //                 setError('Invalid data format received');
+    //             }
+    //         } catch (error) {
+    //             console.error('Error fetching medicines:', error);
+    //             setError('Failed to load medicines');
+    //             setMedicines([]);
+    //         } finally {
+    //             setIsLoading(false);
+    //         }
+    //     };
+
+    //     fetchUser();
+    //     fetchMedicines();
+    // }, [router]);
+
+    // useEffect(() => {
+    //     fetch("api/queryAll")
+    //         .then((response) => response.json())
+    //         .then((data) => {
+    //             const filteredMedicines = data.filter(medicine => medicine.PostingHospital !== loggedInHospital);
+    //             setMedicines(filteredMedicines);
+
+    //             const borrowCount = data.reduce((total, medicine) => {
+    //                 if (!medicine.BorrowRecords) return total;
+    //                 const hospitalBorrows = medicine.BorrowRecords.filter(
+    //                     record => record.BorrowingHospital === loggedInHospital
+    //                 ).length;
+    //                 return total + hospitalBorrows;
+    //             }, 0);
+
+    //             setTotalBorrow(borrowCount);
+    //         })
+    //         .catch((error) => console.error("Error fetching data:", error));
+    // }, [loggedInHospital]);
 
     return (
         <div className="container mx-auto p-4">
