+++ src/middleware.ts
@@ -2,7 +2,7 @@ import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
 import { jwtVerify } from 'jose';
 
-const JWT_SECRET = process.env.NEXT_PUBLIC_JWT_SECRET || 'your-super-secret-key-change-this-in-production';
+const JWT_SECRET = process.env.JWT_SECRET || 'your-super-secret-key-change-this-in-production';
 
 // This function can be marked `async` if using `await` inside
 export async function middleware(request: NextRequest) {
@@ -10,11 +10,11 @@ export async function middleware(request: NextRequest) {
   const path = request.nextUrl.pathname;
 
   // Define public paths that don't require authentication
-  const isPublicPath = path === '/' || path === '/register' || path === '/api/auth/login' || path === '/api/auth/register';
+  const isPublicPath = path === '/' || path === '/register' || path === '/api/auth/login' || path === '/api/auth/register' || path === '/login';
 
   // Get the token from the cookies
   const token = request.cookies.get('token')?.value || '';
-
+  
   // Verify the token
   let isAuthenticated = false;
   if (token) {
